requirements.txt
openai
python-dotenv
chromadb
pypdf
lxml
ipykernel
langchain
langchain-community
langchain-text-splitters
langchain-openai
sentence-transformers
langchain_huggingface
langgraph
faiss-cpu
flask
bs4
uvicorn
PyPDF2
python-multipart
fastapi[all]
python-dotenv
openai
requests
flask
langchain
langchain-community
langchain-openai
langchain-chroma
langchain-text-splitters
chromadb
pypdf
PyPDF2
ipykernel
uvicorn
prompty 
azure-identity 
azure-cosmos
azure-ai-openai